Fixed a bug that prevented fractional hz values from working correctly.
Detail: The program tries to allow the user to alter each transmission event in a given period.  However, if the user specified "-i 0.666" or "-f 1.5" the program will now correctly simplify this to a single transmission event.  If the user wanted to treat it as more, they could alter the period.  For example "-f 3 -P 2" would have a similar interval between packets (0.666 s) and allow control of 3 different transmission events.